refactor: detail 페이지 컴포넌트 분리 및 커스텀 훅 추가
