+++ src/hooks/detail/useArticleRotation.ts
@@ -1,0 +1,19 @@
+import { useState, useEffect } from 'react';
+import { DetailArticleData } from '@/types';
+
+export function useArticleRotation(articles: DetailArticleData[] | undefined) {
+  const [currentIndex, setCurrentIndex] = useState(0);
+
+  useEffect(() => {
+    const articleCount = articles?.length || 0;
+    if (articleCount > 1) {
+      const interval = setInterval(() => {
+        setCurrentIndex((prev) => (prev + 1) % articleCount);
+      }, 15000);
+
+      return () => clearInterval(interval);
+    }
+  }, [articles]);
+
+  return { currentIndex };
+}
