+++ src/pages/Detail/index.tsx
@@ -1,48 +1,41 @@
-import { Text } from '@/components/common/typography/Text';
-import { BsFillQuestionCircleFill } from 'react-icons/bs';
-import { IoMdRefresh } from 'react-icons/io';
-import TickerCharts from '@/components/Ticker/TickerCharts';
-import RealTimeTickerCharts from '@/components/Ticker/RealTimeTickerCharts';
-import ScoreGaugeChart from '@/components/Detail/ScoreGaugeChart';
-import Button from '@/components/common/Button';
+import { useState, useEffect, useCallback, useRef } from 'react';
 import { useParams, useLocation } from 'react-router-dom';
 import styled from 'styled-components';
+
 import { useGetTickerDetail } from '@/api/hooks/useGetTickerDetail';
 import { useGetRealGraph, RealGraphPeriod } from '@/api/hooks/useGetRealGraph';
 import { useGetRealTimePrice } from '@/api/hooks/useGetRealTimePrice';
-import { useGetRealTimeStream } from '@/api/hooks/useGetRealTimeStream';
 import { useGetArticleSummaryTicker } from '@/api/hooks/useGetArticleSummaryTicker';
-import { usePutFavoriteTicker } from '@/api/hooks/usePutFavoriteTicker';
-import Loading from '@/components/common/Layout/Loading';
-import { TickerRealTimeGraphResponse } from '@/types';
-import { useState, useEffect, useRef, useCallback } from 'react';
 import { useQueryClient } from '@tanstack/react-query';
-import { Paragraph } from '@/components/common/typography/Paragraph';
-import RotationArticleItem from '@/components/Detail/RotationArticleItem';
-import SummaryModal from '@/components/Detail/SummaryModal';
+
+import { usePutFavoriteTicker } from '@/api/hooks/usePutFavoriteTicker';
+import { useLiveChart } from '@/hooks/detail/useLiveChart';
+import { useArticleRotation } from '@/hooks/detail/useArticleRotation';
 import useIsMobile from '@/hooks/useIsMobile';
 import useAuth from '@/hooks/useAuth';
-import { MdOutlineStickyNote2 } from 'react-icons/md';
-import { PiHeartFill, PiHeartLight } from 'react-icons/pi';
+
+import Loading from '@/components/common/Layout/Loading';
+import SummaryModal from '@/components/Detail/SummaryModal';
 import LoginModal from '@/components/common/Modal/LoginModal';
+import TickerHeader from '@/components/Detail/TickerHeader';
+import TickerPriceSection from '@/components/Detail/TickerPriceSection';
+import ChartSection from '@/components/Detail/ChartSection';
+import ArticleSection from '@/components/Detail/ArticleSection';
 
 export default function DetailPage() {
   const { id } = useParams() as { id: string };
-  const [period, setPeriod] = useState<RealGraphPeriod>('2W');
-  const [isLiveMode, setIsLiveMode] = useState(false);
-  const [liveChartData, setLiveChartData] = useState<
-    TickerRealTimeGraphResponse[]
-  >([]);
-  const [showTooltip, setShowTooltip] = useState(false);
-  const [showRefreshTooltip, setShowRefreshTooltip] = useState(false);
-  const [currentNewsIndex, setCurrentNewsIndex] = useState(0);
-  const [showSummaryModal, setShowSummaryModal] = useState(false);
-  const [showLoginModal, setShowLoginModal] = useState(false);
   const location = useLocation();
-  const scrollPositionRef = useRef(0);
-  const queryClient = useQueryClient();
   const isMobile = useIsMobile();
   const { isAuthenticated } = useAuth();
+
+  const [showSummaryModal, setShowSummaryModal] = useState(false);
+  const [showLoginModal, setShowLoginModal] = useState(false);
+
+  const [period, setPeriod] = useState<RealGraphPeriod>('2W');
+  const [isLiveMode, setIsLiveMode] = useState(false);
+  const [showRefreshTooltip, setShowRefreshTooltip] = useState(false);
+  const scrollPositionRef = useRef(0);
+  const queryClient = useQueryClient();
 
   const {
     data: tickerResponse,
@@ -53,26 +46,22 @@ export default function DetailPage() {
     data: realGraphResponse,
     isLoading: realGraphLoading,
     error: realGraphError,
-  } = useGetRealGraph({
-    tickerId: id,
-    period,
-  });
+  } = useGetRealGraph({ tickerId: id, period });
   const {
     data: realTimePriceResponse,
     isLoading: realTimePriceLoading,
     error: realTimePriceError,
-  } = useGetRealTimePrice({
-    tickerId: id,
-    enabled: isAuthenticated,
-  });
+  } = useGetRealTimePrice({ tickerId: id, enabled: isAuthenticated });
   const { data: summaryResponse } = useGetArticleSummaryTicker(id);
-  const { mutate: putFavoriteTicker } = usePutFavoriteTicker();
 
   const tickerData = tickerResponse?.content;
   const realGraphData = realGraphResponse?.content;
   const realTimePriceData = realTimePriceResponse?.content;
   const summaryData = summaryResponse?.content ?? null;
+  const articles = tickerData?.detailData.article;
+
   const [isFavorite, setIsFavorite] = useState(false);
+  const { mutate: putFavoriteTicker } = usePutFavoriteTicker();
 
   useEffect(() => {
     if (tickerData?.isFavorite !== undefined) {
@@ -105,51 +94,16 @@ export default function DetailPage() {
     },
     [isFavorite, isAuthenticated, tickerData?.tickerCode, putFavoriteTicker]
   );
-
-  useEffect(() => {
-    if (realTimePriceData?.priceDataList) {
-      setLiveChartData(realTimePriceData.priceDataList);
-    }
-  }, [realTimePriceData]);
-
-  useGetRealTimeStream(id, isLiveMode && isAuthenticated, (newItem) => {
-    setLiveChartData((prev) => {
-      const last = prev[prev.length - 1];
-      const toMinute = (time: string) => time.slice(0, 5);
-
-      // 같은 1분봉이면 마지막 가격만 실시간 업데이트
-      if (last && toMinute(last.hours) === toMinute(newItem.time)) {
-        return [
-          ...prev.slice(0, -1),
-          {
-            ...last,
-            price: newItem.close,
-            hours: newItem.time,
-          },
-        ];
-      }
-
-      const nextIndex = (last?.index ?? -1) + 1;
-      return [
-        ...prev,
-        { price: newItem.close, hours: newItem.time, index: nextIndex },
-      ];
-    });
-  });
+  const { liveChartData } = useLiveChart(
+    id,
+    isLiveMode,
+    isAuthenticated,
+    realTimePriceData
+  );
+  const { currentIndex } = useArticleRotation(articles);
 
   const isLoading = tickerLoading || realGraphLoading || realTimePriceLoading;
   const error = tickerError || realGraphError || realTimePriceError;
-
-  useEffect(() => {
-    const articleCount = tickerData?.detailData.article?.length || 0;
-    if (articleCount > 1) {
-      const interval = setInterval(() => {
-        setCurrentNewsIndex((prevIndex) => (prevIndex + 1) % articleCount);
-      }, 15000);
-
-      return () => clearInterval(interval);
-    }
-  }, [tickerData?.detailData.article?.length]);
 
   const handleRefresh = () => {
     if (isLiveMode) {
@@ -162,9 +116,7 @@ export default function DetailPage() {
       });
     }
     setShowRefreshTooltip(true);
-    setTimeout(() => {
-      setShowRefreshTooltip(false);
-    }, 3000);
+    setTimeout(() => setShowRefreshTooltip(false), 3000);
   };
 
   const handlePeriodChange = (newPeriod: RealGraphPeriod) => {
@@ -191,23 +143,13 @@ export default function DetailPage() {
       });
     });
   };
-
-  const formatDate = (priceDate: string) => {
-    const [, month, day] = priceDate.split('-');
-    return `${Number(month)}월 ${Number(day)}일`;
-  };
-
-  if (isLoading) {
-    return <Loading />;
-  }
-  if (error) {
+  if (isLoading) return <Loading />;
+  if (error)
     return (
       <ErrorMessage>데이터를 불러오는 중 오류가 발생했습니다.</ErrorMessage>
     );
-  }
-  if (!tickerData) {
+  if (!tickerData)
     return <ErrorMessage>주식 정보를 찾을 수 없습니다.</ErrorMessage>;
-  }
 
   return (
     <>
@@ -217,235 +159,34 @@ export default function DetailPage() {
           onClose={() => setShowSummaryModal(false)}
           summaryData={summaryData}
         />
-        <TickerTitle>
-          <CompanyInfo>
-            <Text size={isMobile ? 'm' : 'l'} weight="bold">
-              {tickerData.shortCompanyName}
-            </Text>
-            <Text size={isMobile ? 'xs' : 's'} weight="normal" variant="grey">
-              {tickerData.tickerCode}
-            </Text>
-            <LikeIconWrapper onClick={handleLikeClick}>
-              {isFavorite ? (
-                <PiHeartFill color="#fe7373" size={isMobile ? 18 : 22} />
-              ) : (
-                <PiHeartLight size={isMobile ? 18 : 22} color="#ccc" />
-              )}
-            </LikeIconWrapper>
-          </CompanyInfo>
-          {!isMobile && (
-            <ScoreTitleContainer>
-              <Paragraph size={isMobile ? 'xs' : 's'} weight="bold">
-                종목 점수
-              </Paragraph>
-              <TooltipContainer
-                onMouseEnter={() => setShowTooltip(true)}
-                onMouseLeave={() => setShowTooltip(false)}
-              >
-                <BsFillQuestionCircleFill
-                  size={isMobile ? 14 : 16}
-                  color="#BCC7D9"
-                />
-                {showTooltip && (
-                  <Tooltip>
-                    수집된 기사의 감정(긍정/부정) 비율에 추세를 반영하여 계산된
-                    점수입니다.
-                  </Tooltip>
-                )}
-              </TooltipContainer>
-            </ScoreTitleContainer>
-          )}
-        </TickerTitle>
-        <TickerInfo>
-          <InfoGrid>
-            <InfoItem>
-              <Text size={isMobile ? 'xxs' : 'xs'} weight="normal">
-                시가
-              </Text>
-              <Text
-                size={isMobile ? 'xxs' : 'xs'}
-                weight="normal"
-                variant="grey"
-              >
-                $ {tickerData.detailData.open}
-              </Text>
-            </InfoItem>
-
-            <InfoItem>
-              <Text size={isMobile ? 'xxs' : 'xs'} weight="normal">
-                종가
-              </Text>
-              <Text
-                size={isMobile ? 'xxs' : 'xs'}
-                weight="normal"
-                variant="grey"
-              >
-                $ {tickerData.detailData.close}
-              </Text>
-            </InfoItem>
-
-            <InfoItem>
-              <Text size={isMobile ? 'xxs' : 'xs'} weight="normal">
-                고가
-              </Text>
-              <Text
-                size={isMobile ? 'xxs' : 'xs'}
-                weight="normal"
-                variant="grey"
-              >
-                $ {tickerData.detailData.high}
-              </Text>
-            </InfoItem>
-
-            <InfoItem>
-              <Text size={isMobile ? 'xxs' : 'xs'} weight="normal">
-                저가
-              </Text>
-              <Text
-                size={isMobile ? 'xxs' : 'xs'}
-                weight="normal"
-                variant="grey"
-              >
-                $ {tickerData.detailData.low}
-              </Text>
-            </InfoItem>
-
-            <InfoItem>
-              <Text size={isMobile ? 'xxs' : 'xs'} weight="normal">
-                거래량
-              </Text>
-              <Text
-                size={isMobile ? 'xxs' : 'xs'}
-                weight="normal"
-                variant="grey"
-              >
-                {tickerData.detailData.volume.toLocaleString()}주
-              </Text>
-            </InfoItem>
-            <ItemDate>
-              <Text size="xxs" weight="normal" variant="grey">
-                * {formatDate(tickerData.detailData.priceDate)} 기준
-              </Text>
-            </ItemDate>
-          </InfoGrid>
-          {!isMobile && (
-            <ScoreGaugeChart
-              value={tickerData.sentimentScore}
-              maxValue={100}
-              title="점수"
-              predictionStrategy={tickerData.predictionStrategy}
-            />
-          )}
-        </TickerInfo>
-        {isMobile && (
-          <ScoreTitleContainer>
-            <Paragraph size={isMobile ? 'xs' : 's'} weight="bold">
-              종목 점수
-            </Paragraph>
-            <TooltipContainer
-              onMouseEnter={() => setShowTooltip(true)}
-              onMouseLeave={() => setShowTooltip(false)}
-            >
-              <BsFillQuestionCircleFill
-                size={isMobile ? 14 : 16}
-                color="#BCC7D9"
-              />
-              {showTooltip && (
-                <Tooltip>
-                  수집된 기사의 감정(긍정/부정) 비율에
-                  <br /> 추세를 반영하여 계산된 점수입니다.
-                </Tooltip>
-              )}
-            </TooltipContainer>
-          </ScoreTitleContainer>
-        )}
-        {isMobile && (
-          <ScoreGaugeChart
-            value={tickerData.sentimentScore}
-            maxValue={100}
-            title="점수"
-            predictionStrategy={tickerData.predictionStrategy}
+        <TickerHeader
+          tickerData={tickerData}
+          isMobile={isMobile}
+          isFavorite={isFavorite}
+          onLikeClick={handleLikeClick}
+        />
+        <TickerPriceSection tickerData={tickerData} isMobile={isMobile} />
+        <ChartSection
+          isMobile={isMobile}
+          period={period}
+          isLiveMode={isLiveMode}
+          showRefreshTooltip={showRefreshTooltip}
+          realGraphData={realGraphData}
+          tickerData={tickerData}
+          liveChartData={liveChartData}
+          onRefresh={handleRefresh}
+          onPeriodChange={handlePeriodChange}
+          onLiveMode={handleLiveMode}
+          onShowSummary={() => setShowSummaryModal(true)}
+        />
+        {articles && articles.length > 0 && (
+          <ArticleSection
+            articles={articles}
+            currentIndex={currentIndex}
+            tickerCode={tickerData.tickerCode}
+            isMobile={isMobile}
           />
         )}
-        <StockPriceSection>
-          <Paragraph size={isMobile ? 'xs' : 's'} weight="bold">
-            실제 주가
-          </Paragraph>
-          {isMobile && (
-            <SummaryButton
-              onClick={() => setShowSummaryModal(true)}
-              variant="grey"
-              size="small"
-            >
-              <MdOutlineStickyNote2 size={16} />
-            </SummaryButton>
-          )}
-        </StockPriceSection>
-        <PeriodSelectorContainer>
-          <PeriodSelector>
-            {(['2W', '1M', '6M', '1Y'] as RealGraphPeriod[]).map((p) => (
-              <PeriodButton
-                key={p}
-                $active={period === p && !isLiveMode}
-                onClick={() => handlePeriodChange(p)}
-              >
-                {p}
-              </PeriodButton>
-            ))}
-            <LiveButton $active={isLiveMode} onClick={handleLiveMode}>
-              live
-              <LiveDot />
-            </LiveButton>
-          </PeriodSelector>
-          <ButtonGroup>
-            {!isMobile && (
-              <SummaryButton
-                onClick={() => setShowSummaryModal(true)}
-                variant="grey"
-                size="small"
-              >
-                <MdOutlineStickyNote2 size={16} />
-              </SummaryButton>
-            )}
-            <RefreshContainer>
-              <RefreshButton
-                onClick={handleRefresh}
-                variant="grey"
-                size="small"
-              >
-                <IoMdRefresh size={isMobile ? 14 : 16} />
-              </RefreshButton>
-              {showRefreshTooltip && (
-                <RefreshTooltip>
-                  최신 상태로 업데이트 되었습니다!
-                </RefreshTooltip>
-              )}
-            </RefreshContainer>
-          </ButtonGroup>
-        </PeriodSelectorContainer>
-        {isLiveMode ? (
-          <RealTimeTickerCharts realTimeData={liveChartData} />
-        ) : (
-          realGraphData && (
-            <TickerCharts
-              realData={realGraphData.graphData || []}
-              sentiment={tickerData.sentiment ?? 0}
-            />
-          )
-        )}
-        {tickerData?.detailData.article &&
-          tickerData.detailData.article.length > 0 && (
-            <>
-              <Paragraph size={isMobile ? 'xs' : 's'} weight="bold">
-                실시간 기사
-              </Paragraph>
-              <RotationArticleItem
-                key={currentNewsIndex}
-                item={tickerData.detailData.article[currentNewsIndex]}
-                tickerCode={tickerData.tickerCode}
-              />
-            </>
-          )}
       </Wrapper>
       {showLoginModal && (
         <LoginModal
@@ -457,6 +198,7 @@ export default function DetailPage() {
     </>
   );
 }
+
 const Wrapper = styled.div`
   width: 90%;
   display: flex;
@@ -470,357 +212,9 @@ const Wrapper = styled.div`
     padding: 12px 0;
   }
 `;
-const TickerTitle = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: baseline;
-
-  @media screen and (max-width: 768px) {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 8px;
-  }
-`;
-
-const CompanyInfo = styled.div`
-  display: flex;
-  align-items: baseline;
-  gap: 8px;
-  margin-right: 20px;
-
-  @media screen and (max-width: 768px) {
-    gap: 6px;
-    margin-right: 2px;
-  }
-`;
-
-const ScoreTitleContainer = styled.div`
-  display: flex;
-  gap: 6px;
-  margin-right: 180px;
-
-  @media screen and (max-width: 768px) {
-    margin-right: 10px;
-    gap: 4px;
-  }
-`;
-
-const TooltipContainer = styled.div`
-  position: relative;
-  display: flex;
-`;
-
-const Tooltip = styled.div`
-  position: absolute;
-  bottom: 100%;
-  left: 0;
-  background: white;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  padding: 12px;
-  font-size: 12px;
-  color: #333;
-  white-space: nowrap;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-  margin-bottom: 8px;
-  z-index: 1000;
-
-  &::after {
-    content: '';
-    position: absolute;
-    top: 100%;
-    left: 12px;
-    border: 6px solid transparent;
-    border-top-color: white;
-  }
-
-  &::before {
-    content: '';
-    position: absolute;
-    top: 100%;
-    left: 12px;
-    border: 7px solid transparent;
-    border-top-color: #ddd;
-    margin-top: 1px;
-  }
-
-  @media screen and (max-width: 768px) {
-    line-height: 1.4;
-  }
-`;
-const TickerInfo = styled.div`
-  display: flex;
-  align-items: flex-start;
-  gap: 20px;
-  padding: 8px 0px 12px 0px;
-
-  @media screen and (max-width: 768px) {
-    flex-direction: column;
-    gap: 16px;
-  }
-`;
-const InfoGrid = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  width: 340px;
-  border-radius: 8px;
-  gap: 24px;
-  padding: 24px 0 0 0;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-
-  @media screen and (max-width: 768px) {
-    width: 90%;
-    max-width: 300px;
-    gap: 20px;
-    padding: 16px 0 0 0;
-  }
-`;
-
-const InfoItem = styled.div`
-  display: flex;
-  align-items: flex-end;
-  justify-content: space-between;
-  margin: 0px 26px;
-
-  &:nth-child(5) {
-    grid-column: span 2;
-    justify-content: flex-start;
-    gap: 24px;
-    margin: 0 0 0 26px;
-  }
-
-  &:nth-child(6) {
-    justify-content: flex-end;
-  }
-
-  @media screen and (max-width: 768px) {
-    margin: 0px 20px;
-
-    &:nth-child(5) {
-      gap: 16px;
-      margin: 0 0 0 20px;
-    }
-  }
-`;
-const ItemDate = styled.div`
-  grid-column: span 2;
-  text-align: right;
-  margin-right: 26px;
-  margin-top: -40px;
-
-  @media screen and (max-width: 768px) {
-    margin-right: 20px;
-    margin-top: -32px;
-  }
-`;
-const PeriodSelector = styled.div`
-  display: flex;
-  gap: 12px;
-
-  @media screen and (max-width: 768px) {
-    gap: 8px;
-  }
-`;
-const PeriodButton = styled.button<{ $active: boolean }>`
-  padding: 8px 16px;
-  border: 1px solid ${(props) => (props.$active ? '#47c8d9' : '#ddd')};
-  background-color: ${(props) => (props.$active ? '#47c8d9' : 'white')};
-  color: ${(props) => (props.$active ? 'white' : '#666')};
-  border-radius: 6px;
-  cursor: pointer;
-  font-size: 14px;
-  font-weight: 500;
-  transition: all 0.2s ease;
-
-  &:hover {
-    border-color: #47c8d9;
-    background-color: ${(props) => (props.$active ? '#47c8d9' : '#f0f9fa')};
-  }
-
-  @media screen and (max-width: 768px) {
-    padding: 6px 12px;
-    font-size: 12px;
-  }
-`;
-
-const LiveButton = styled.button<{ $active: boolean }>`
-  display: flex;
-  align-items: center;
-  position: relative;
-  padding: 8px 8px 8px 12px;
-  border: 1px solid ${(props) => (props.$active ? '#47c8d9' : '#ddd')};
-  background-color: ${(props) => (props.$active ? '#47c8d9' : 'white')};
-  color: ${(props) => (props.$active ? 'white' : '#666')};
-  border-radius: 6px;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  gap: 2px;
-
-  &:hover {
-    border-color: #47c8d9;
-    background-color: ${(props) => (props.$active ? '#47c8d9' : '#f0f9fa')};
-  }
-
-  @media screen and (max-width: 768px) {
-    padding: 6px 6px 6px 12px;
-    font-size: 12px;
-  }
-`;
-
-const LiveDot = styled.div`
-  width: 6px;
-  height: 6px;
-  position: relative;
-  top: -8px;
-  background-color: #e74c3c;
-  border-radius: 50%;
-  animation: pulse 2s infinite;
-
-  @media screen and (max-width: 768px) {
-    width: 4px;
-    height: 4px;
-    top: -6px;
-  }
-
-  @keyframes pulse {
-    0% {
-      transform: scale(0.8);
-      opacity: 1;
-    }
-    50% {
-      transform: scale(1.2);
-      opacity: 0.8;
-    }
-    100% {
-      transform: scale(0.8);
-      opacity: 1;
-    }
-  }
-`;
-const StockPriceSection = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 10px;
-`;
-
-const PeriodSelectorContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const ButtonGroup = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  margin-right: 24px;
-
-  @media screen and (max-width: 768px) {
-    gap: 8px;
-    margin-right: 0;
-  }
-`;
-
-const SummaryButton = styled(Button)`
-  width: auto;
-  height: 34px;
-  padding: 0 12px;
-
-  @media screen and (max-width: 768px) {
-    height: 24px;
-    padding: 0 8px;
-  }
-`;
-
-const RefreshContainer = styled.div`
-  position: relative;
-  display: flex;
-`;
-
-const RefreshButton = styled(Button)`
-  width: 40px;
-  height: 34px;
-
-  &:hover {
-    svg {
-      transform: rotate(60deg);
-      transition: transform 0.3s ease;
-    }
-  }
-
-  @media screen and (max-width: 768px) {
-    width: 34px;
-    height: 28px;
-  }
-`;
-
-const RefreshTooltip = styled(Tooltip)`
-  padding: 10px;
-  right: 72px;
-  left: auto;
-  top: 0px;
-  bottom: auto;
-  animation: tooltipSlideLeft 3s ease-in-out forwards;
-
-  &::after,
-  &::before {
-    display: none;
-  }
-
-  @media screen and (max-width: 768px) {
-    right: 24px;
-    top: 34px;
-    animation: tooltipSlideBottom 3s ease-in-out forwards;
-
-    @keyframes tooltipSlideLeft {
-      0% {
-        opacity: 0;
-        transform: translateX(20px);
-      }
-      15% {
-        opacity: 1;
-        transform: translateX(0);
-      }
-      85% {
-        opacity: 1;
-        transform: translateX(0);
-      }
-      100% {
-        opacity: 0;
-        transform: translateX(-20px);
-      }
-    }
-
-    @keyframes tooltipSlideBottom {
-      0% {
-        opacity: 0;
-        transform: translateY(-20px);
-      }
-      15% {
-        opacity: 1;
-        transform: translateY(0);
-      }
-      85% {
-        opacity: 1;
-        transform: translateY(0);
-      }
-      100% {
-        opacity: 0;
-        transform: translateY(20px);
-      }
-    }
-  }
-`;
 
 const ErrorMessage = styled.div`
   padding: 20px;
   text-align: center;
   color: #e74c3c;
-`;
-
-const LikeIconWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  line-height: 1;
 `;
